requirements.txt
anthropic
jugaad-trader
apscheduler
streamlit
pandas
python-dotenv
cryptography
yfinance
fastapi
uvicorn[standard]
python-jose[cryptography]
